Fixed the rest of AboutFunctions.js
The issue was in the white spaces. For starters there was a \u000d
just before the new line. It looks very much like a space, but it
is not one. Also, the leading whitespace for all but the first line
was odd, but finally I figured it out.

Lesson: Don't just trust that a character is what it looks like,
check the unicode value for it.